fix: session table, ROUND cast, hpp view req ref, sequence resync
Four defects surfaced by the first end-to-end run against a real Supabase
Postgres instance. All four were invisible to static review and to the
existing test suite.

- db.js: provision the connect-pg-simple `session` table in
  bootstrapSchema(). The store is constructed without createTableIfMissing,
  so it never creates its own table and every login failed with
  `relation "session" does not exist` (42P01).
- routes/dashboard.js: ROUND(double precision, integer) has no Postgres
  overload; cast the AVG to numeric. SQLite accepted the two-arg form.
- views/hpp/index.ejs: the template referenced `req.query.variant_id`,
  which is not in the EJS locals — use the `selectedVariantId` local the
  route already passes. Pre-existing bug, only reachable once /hpp rendered.
- tmp/migrate-data.js: resync every SERIAL sequence to MAX(id) after
  import. Rows carried their original SQLite ids, which does not advance a
  sequence, so the next app-side INSERT collided on the primary key.
- tmp/resync-sequences.js: standalone idempotent version of that resync,
  for databases already imported without it.
- tests/setup.js: truncate `session` between tests along with the rest.

Verified against Supabase (pooler, ap-southeast-1): jest 3/3 pass,
tmp/smoke-test.js 7/7 pass, and a manual pass over every page plus vendor
CRUD, PO create/approve with FIFO batch + stock movement, production batch
create/cost/delivery/approve with HPP recalculation, and import
create/approve. Zero server-side errors. Smoke data was removed afterward.

+++ routes/dashboard.js
@@ -20,7 +20,7 @@ router.get('/', isAuthenticated, async (req, res) => {
   const totalPending = pendingPO + pendingCost + pendingImport;
 
   const products = (await db.query(`
-    SELECT p.nama_produk, ROUND(AVG(pv.hpp_saat_ini),0) as avg_hpp
+    SELECT p.nama_produk, ROUND(AVG(pv.hpp_saat_ini)::numeric, 0) as avg_hpp
     FROM products p JOIN product_variants pv ON p.id = pv.product_id
     WHERE pv.hpp_saat_ini > 0 GROUP BY p.id
   `)).rows;
